修改了Comment和Discussion表，增加了imgUrl，mentionedUser,topic属性，并修改了对应的实体类，映射接口和Controller

+++ src/main/java/com/example/bcp/controller/DiscussionController.java
@@ -37,6 +37,9 @@ DiscussionController {
             discussionInfo.put("Information",d.getDiscussionInformation());
             discussionInfo.put("PostingTime",d.getDiscussionPostingTime());
             discussionInfo.put("studentNo",d.getStudentNo());
+            discussionInfo.put("imgUrl",d.getImageUrl());
+            discussionInfo.put("mentionedUser",d.getMentionedUser());
+            discussionInfo.put("topic",d.getTopic());
             responseData.put("discussion" + i,discussionInfo);
             i++;
         }
@@ -59,6 +62,9 @@ DiscussionController {
             discussionInfo.put("Information",d.getDiscussionInformation());
             discussionInfo.put("PostingTime",d.getDiscussionPostingTime());
             discussionInfo.put("studentNo",d.getStudentNo());
+            discussionInfo.put("imgUrl",d.getImageUrl());
+            discussionInfo.put("mentionedUser",d.getMentionedUser());
+            discussionInfo.put("topic",d.getTopic());
             responseData.put("discussion" + i,discussionInfo);
             i++;
         }
@@ -66,11 +72,94 @@ DiscussionController {
         return Result.success(responseData);
     }
 
+    /**
+     *  查找指定课程下某一话题的所有帖子
+     */
+    @GetMapping(value = "/topic")
+    public Result getDiscussionWithTopic(@RequestParam String cid,
+                                         @RequestParam String topic,
+                                         HttpServletRequest request){
+        List<Discussion> discussions = discussionService.selectByCidAndTopic(topic,cid);
+        Map<String, Object> responseData = new HashMap<>();
+        int i = 1;
+        for(Discussion d : discussions){
+            Map<String , Object> discussionInfo = new HashMap<>();
+            discussionInfo.put("index",i);
+            discussionInfo.put("DiscussionId",d.getDiscussionId());
+            discussionInfo.put("Information",d.getDiscussionInformation());
+            discussionInfo.put("PostingTime",d.getDiscussionPostingTime());
+            discussionInfo.put("studentNo",d.getStudentNo());
+            discussionInfo.put("imgUrl",d.getImageUrl());
+            discussionInfo.put("mentionedUser",d.getMentionedUser());
+            discussionInfo.put("topic",d.getTopic());
+            responseData.put("discussion" + i,discussionInfo);
+            i++;
+        }
+        return Result.success(responseData);
+    }
+
+    /**
+     *  查找指定课程下某一课程下，被别人@的所有帖子
+     */
+    @GetMapping(value = "/cid_mentioned")
+    public Result getMentionedDiscussionWithCid(@RequestParam String cid,
+                                                HttpServletRequest request){
+        String username = request.getAttribute("username").toString();
+        List<Discussion> discussions = discussionService.selectByCidAndMentioned(username,cid);
+        Map<String, Object> responseData = new HashMap<>();
+        int i = 1;
+        for(Discussion d : discussions){
+            Map<String , Object> discussionInfo = new HashMap<>();
+            discussionInfo.put("index",i);
+            discussionInfo.put("DiscussionId",d.getDiscussionId());
+            discussionInfo.put("Information",d.getDiscussionInformation());
+            discussionInfo.put("PostingTime",d.getDiscussionPostingTime());
+            discussionInfo.put("studentNo",d.getStudentNo());
+            discussionInfo.put("imgUrl",d.getImageUrl());
+            discussionInfo.put("mentionedUser",d.getMentionedUser());
+            discussionInfo.put("topic",d.getTopic());
+            responseData.put("discussion" + i,discussionInfo);
+            i++;
+        }
+        return Result.success(responseData);
+    }
+
+    /**
+     *  查找所有被别人@的所有帖子
+     */
+    @GetMapping("/mentioned")
+    public Result getMentionedDiscussion(HttpServletRequest request){
+        String username = request.getAttribute("username").toString();
+        List<Discussion> discussions = discussionService.selectByMentioned(username);
+        Map<String, Object> responseData = new HashMap<>();
+        int i = 1;
+        for(Discussion d : discussions){
+            Map<String , Object> discussionInfo = new HashMap<>();
+            discussionInfo.put("index",i);
+            discussionInfo.put("DiscussionId",d.getDiscussionId());
+            discussionInfo.put("Information",d.getDiscussionInformation());
+            discussionInfo.put("PostingTime",d.getDiscussionPostingTime());
+            discussionInfo.put("studentNo",d.getStudentNo());
+            discussionInfo.put("imgUrl",d.getImageUrl());
+            discussionInfo.put("mentionedUser",d.getMentionedUser());
+            discussionInfo.put("topic",d.getTopic());
+            responseData.put("discussion" + i,discussionInfo);
+            i++;
+        }
+        return Result.success(responseData);
+    }
+
+    /**
+     *  新增帖子
+     */
     @RequestMapping(value = "/insert" , method = RequestMethod.POST)
     public Result insertDiscussion(@RequestBody Map<String,String> requestData){
         String cid = requestData.get("cid");
         String StudentNo = requestData.get("studentNo");
         String DiscussionInfo = requestData.get("information");
+        String mentionedUser = requestData.get("mentionedUser");
+        String imgUrl = requestData.get("imgUrl");
+        String tag = requestData.get("tag");
         LocalDateTime postingTime  = LocalDateTime.now();
         int size = discussionService.selectAllDiscussion().size()+1;
         String DiscussionId = "D"+size;
@@ -83,9 +172,10 @@ DiscussionController {
             }
         }
         String message ;
+        System.out.println(DiscussionId);
         try{
             discussionService.insertDiscussion(StudentNo,cid,DiscussionId,
-                    DiscussionInfo,postingTime);
+                    DiscussionInfo,postingTime,mentionedUser,imgUrl,tag);
             message = "帖子新建成功！";
         }catch (Exception e){
             message = "帖子创建失败，请重试";
@@ -94,20 +184,33 @@ DiscussionController {
         return Result.success(message);
     }
 
+    /**
+     *  删除帖子
+     */
     @RequestMapping(value = "/delete" , method = RequestMethod.POST)
-    public Result deleteDiscussion(@RequestBody Map<String,String> requestData){
-        String DiscussionId =requestData.get("discussionId");
-        String message ;
-        try{
-            discussionService.deleteDiscussion(DiscussionId);
-            message = "帖子删除成功！";
-        }catch (Exception e){
-            message = "帖子删除失败，请重试";
-            return Result.error(message);
-        }
-        return Result.success(message);
+    public Result deleteDiscussion(@RequestBody Map<String,String> requestData,HttpServletRequest request){
+       String username = request.getAttribute("username").toString();
+       if (username.startsWith("T")){
+           String DiscussionId =requestData.get("discussionId");
+           String message ;
+           try{
+               discussionService.deleteDiscussion(DiscussionId);
+               message = "帖子删除成功！";
+           }catch (Exception e){
+               message = "帖子删除失败，请重试";
+               return Result.error(message);
+           }
+           return Result.success(message);
+       }else {
+           String log = "无删除权限";
+           return Result.error(log);
+       }
+
     }
 
+    /**
+     *  修改帖子的信息
+     */
     @RequestMapping(value = "/updateInfo" , method = RequestMethod.POST)
     public Result updateDiscussionInfo(@RequestBody Map<String,String> requestData){
         String DiscussionInfo = requestData.get("discussionInfo");
